refactor: homescreen icon handling
Only store the image filename instead of the full file path.
This allows for a future migration of the downloaded metadata. The
`UC_DATA_HOME` location cannot be considered static.
Furthermore, using relative paths only worked if the `UC_DATA_HOME` was
the working directory. This is not the case on the Remote device.