refactor: change metadata icon sizes to 240x240
Most app icons look pretty good with this resolution while keeping a
small size. Since the app icons are encoded as base64 data and sent over
WebSockets, it's important to keep the size as small as possible.

Once we allow file sharing from an integration, the UI can directly
access the files and the image size can be increased up to 480x480.