refactor: metadata icon data handling (#66)
* refactor: homescreen icon handling

Only store the image filename instead of the full file path.
This allows for a future migration of the downloaded metadata. The
`UC_DATA_HOME` location cannot be considered static.
Furthermore, using relative paths only worked if the `UC_DATA_HOME` was
the working directory. This is not the case on the Remote device.

* docs: label Cast & Google Play metadata as preview features

* Add missing French & German translations

* build: update ucapi to 0.3.0

* refactor: empty data folder as default

The cache file is automatically created and doesn't need to exist at
driver startup.
This allows an easier installation on the Remote device and simply
wiping the data cache directory in the future to remove the cache.

* refactor: filter out base64 encoded images in log messages

Add initial unit tests and rename the intg-androidtv source directory
to src. This was required to make the unit tests work and also follows
common project setup practices.

* refactor: change metadata icon sizes to 240x240

Most app icons look pretty good with this resolution while keeping a
small size. Since the app icons are encoded as base64 data and sent over
WebSockets, it's important to keep the size as small as possible.

Once we allow file sharing from an integration, the UI can directly
access the files and the image size can be increased up to 480x480.

* docs: update changelog

* refactor: Chromecast connection state logging